fix: [Scrap] 스크랩 취소 기능 수정
- userId, noticeId 로 조회하던 것을 scrapId로 조회하도록 수정

+++ src/main/java/com/bcsdlab/biseo/serviceImpl/ScrapServiceImpl.java
@@ -40,7 +40,7 @@ public class ScrapServiceImpl implements ScrapService {
         if (scrap != null) {
             throw new RuntimeException("스크랩이 이미 존재합니다.");
         }
-        
+
         ScrapModel scrapModel = new ScrapModel();
         scrapModel.setUserId(user.getId());
         scrapModel.setNoticeId(noticeId);
@@ -50,15 +50,19 @@ public class ScrapServiceImpl implements ScrapService {
     }
 
     @Override
-    public void deleteScrap(Long noticeId) {
+    public void deleteScrap(Long scrapId) {
         UserModel user = userRepository.findById(Long.parseLong(jwtUtil.findUserInfoInToken().getAudience().get(0)));
         if (user == null) {
             throw new RuntimeException("유저가 존재하지 않습니다.");
         }
 
-        ScrapModel scrap = scrapRepository.findScrapByUserIdAndNoticeId(user.getId(), noticeId);
+        ScrapModel scrap = scrapRepository.findScrapById(scrapId);
         if (scrap == null) {
             throw new RuntimeException("스크랩이 존재하지 않습니다.");
+        }
+
+        if (!scrap.getUserId().equals(user.getId())) {
+            throw new RuntimeException("스크랩 삭제 권한이 없습니다.");
         }
 
         scrapRepository.deleteScrapById(scrap.getId());
